requirements.txt
requests
python-dateutil
gitpython